// crates/sdk-kit/src/types.ts
export type ScenarioOverride = Readonly<{
    id: string;
    templateId: string;
    values: Record<string, unknown>;
    scenarioRelativeSlot: number;
    label: string;
    enabled: boolean;
    fetchBeforeUse: boolean;
    account: Readonly<{ pubkey: Address }>;
}>;

export type Scenario = Readonly<{
    id: string;
    name: string;
    description: string;
    overrides: readonly ScenarioOverride[];
    tags: readonly string[];
}>;